backend/pyproject.toml
[project]
name = "blockme-backend"
version = "0.1.0"
description = "BlockMe Backend - FastAPI service for document processing and tax Q&A"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    # AI API 客户端
    "anthropic>=0.40.0",
    "zhipuai>=2.1.0",
    "openai>=1.50.0",

    # 文档处理
    "pdf2image>=1.17.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",

    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",

    # 工具库
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

semantic-cache = [
    "sentence-transformers>=2.7.0",  # 用于语义缓存（任务17）
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
